requirements.txt
paho-mqtt
Pillow
pymongo
requests
Gradio
tenacity